Change type of buff parameter to void *
This commit changes the buff parameter of the CRC functions to void*.
This makes the function easier to use as this way the compiler will
accept any pointer type and convert it to void * without casting.

This means we can simply type:
    struct Foo { ... } bar;
    CRC::crc16(&bar, sizeof(bar));

Instead of:
    struct Foo { ... } bar;
    CRC::crc16((void *)(&bar), sizeof(bar));

Please note that void pointers are also used by standard C library
functions, which work with memory buffers (e.g. memcpy, memset, etc.).